Fix health reporting
