+++ src/Stryker.Core/Stryker.Core.UnitTest/Reporters/JsonReporterTests.cs
@@ -1,18 +1,13 @@
 ﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.Extensions.Logging;
 using Moq;
 using Shouldly;
 using Stryker.Core.Logging;
-using Stryker.Core.Mutants;
-using Stryker.Core.Mutators;
 using Stryker.Core.Options;
 using Stryker.Core.ProjectComponents;
 using Stryker.Core.Reporters.Json;
 using System;
-using System.Collections.ObjectModel;
 using System.IO;
 using System.IO.Abstractions.TestingHelpers;
 using System.Linq;
@@ -61,7 +56,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReportFileComponent_ShouldHaveLanguageSetToCs()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
             var fileComponent = (folderComponent as FolderComposite).GetAllFiles().First();
 
             new JsonReportFileComponent(fileComponent).Language.ShouldBe("cs");
@@ -70,7 +65,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReportFileComponent_ShouldContainOriginalSource()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
             var fileComponent = (folderComponent as FolderComposite).GetAllFiles().First();
 
             new JsonReportFileComponent(fileComponent).Source.ShouldBe(fileComponent.SourceCode);
@@ -79,7 +74,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReportFileComponents_ShouldContainMutants()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
             foreach (var file in (folderComponent as FolderComposite).GetAllFiles())
             {
                 var jsonReportComponent = new JsonReportFileComponent(file);
@@ -94,7 +89,7 @@ namespace Stryker.Core.UnitTest.Reporters
         public void JsonReportFileComponent_DoesNotContainDuplicateMutants()
         {
             var loggerMock = Mock.Of<ILogger>();
-            var folderComponent = CreateProjectWith(duplicateMutant: true);
+            var folderComponent = JsonReportTestHelper.CreateProjectWith(duplicateMutant: true);
             foreach (var file in (folderComponent as FolderComposite).GetAllFiles())
             {
                 var jsonReportComponent = new JsonReportFileComponent(file, loggerMock);
@@ -108,7 +103,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReport_ThresholdsAreSet()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
 
             var report = JsonReport.Build(new StrykerOptions(), folderComponent);
 
@@ -120,7 +115,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReport_ShouldContainAtLeastOneFile()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
 
             var report = JsonReport.Build(new StrykerOptions(), folderComponent);
 
@@ -130,7 +125,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReport_WithMutationScoreOverThresholdHighHasGoodHealth()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
 
             var report = JsonReport.Build(new StrykerOptions(thresholdHigh: 20, thresholdLow: 10), folderComponent);
 
@@ -138,41 +133,41 @@ namespace Stryker.Core.UnitTest.Reporters
         }
 
         [Fact]
-        public void JsonReport_WithMutationScoreEqualToThresholdHighHasGoodHealth()
+        public void JsonReport_WithMutationScoreEqualToThresholdHighHasWarningHealth()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
 
-            var report = JsonReport.Build(new StrykerOptions(thresholdHigh: 66, thresholdLow: 10), folderComponent);
-
-            report.Files.First().Value.Health.ShouldBe(Health.Good);
-        }
-
-        [Fact]
-        public void JsonReport_WithMutationScoreBetweenThresholdHighAndBreakHasWarningHealth()
-        {
-            var folderComponent = CreateProjectWith();
-
-            var report = JsonReport.Build(new StrykerOptions(thresholdHigh: 80, thresholdBreak: 50), folderComponent);
+            var report = JsonReport.Build(new StrykerOptions(thresholdHigh: 67, thresholdLow: 10), folderComponent);
 
             report.Files.First().Value.Health.ShouldBe(Health.Warning);
         }
 
         [Fact]
-        public void JsonReport_WithMutationScoreUnderThresholdBreakHasDangerHealth()
+        public void JsonReport_WithMutationScoreBetweenThresholdHighInclusiveAndLowNonInclusiveHasWarningHealth()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
 
-            var report = JsonReport.Build(new StrykerOptions(thresholdHigh: 80, thresholdLow: 70, thresholdBreak: 67), folderComponent);
+            var report = JsonReport.Build(new StrykerOptions(thresholdHigh: 67, thresholdLow: 66), folderComponent);
 
-            report.Files.First().Value.Health.ShouldBe(Health.Danger);
+            report.Files.First().Value.Health.ShouldBe(Health.Warning);
         }
 
         [Fact]
-        public void JsonReport_WithMutationScoreEqualToThresholdBreakHasDangerHealth()
+        public void JsonReport_WithMutationScoreEqualToThresholdLowHasWarningHealth()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
 
-            var report = JsonReport.Build(new StrykerOptions(thresholdHigh: 80, thresholdLow: 70, thresholdBreak: 66), folderComponent);
+            var report = JsonReport.Build(new StrykerOptions(thresholdHigh: 80, thresholdLow: 65), folderComponent);
+
+            report.Files.First().Value.Health.ShouldBe(Health.Warning);
+        }
+
+        [Fact]
+        public void JsonReport_WithMutationScoreUnderThresholdLowHasDangerHealth()
+        {
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
+
+            var report = JsonReport.Build(new StrykerOptions(thresholdHigh: 80, thresholdLow: 67), folderComponent);
 
             report.Files.First().Value.Health.ShouldBe(Health.Danger);
         }
@@ -180,7 +175,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReport_BuildReportReturnsSingletonJsonReport()
         {
-            var folderComponent = CreateProjectWith();
+            var folderComponent = JsonReportTestHelper.CreateProjectWith();
             var options = new StrykerOptions();
 
             var firstReport = JsonReport.Build(options, folderComponent);
@@ -196,55 +191,9 @@ namespace Stryker.Core.UnitTest.Reporters
             var options = new StrykerOptions(thresholdBreak: 0, thresholdHigh: 80, thresholdLow: 60);
             var reporter = new JsonReporter(options, mockFileSystem);
 
-            reporter.OnAllMutantsTested(CreateProjectWith());
+            reporter.OnAllMutantsTested(JsonReportTestHelper.CreateProjectWith());
             var reportPath = Path.Combine(options.OutputPath, "reports", $"mutation-report.json");
             mockFileSystem.FileExists(reportPath).ShouldBeTrue($"Path {reportPath} should exist but it does not.");
-        }
-
-        private IReadOnlyInputComponent CreateProjectWith(bool duplicateMutant = false, int mutationScore = 60)
-        {
-            var tree = CSharpSyntaxTree.ParseText("void M(){ int i = 0 + 8; }");
-            var originalNode = tree.GetRoot().DescendantNodes().OfType<BinaryExpressionSyntax>().First();
-
-            var mutation = new Mutation()
-            {
-                OriginalNode = originalNode,
-                ReplacementNode = SyntaxFactory.BinaryExpression(SyntaxKind.SubtractExpression, originalNode.Left, originalNode.Right),
-                DisplayName = "This name should display",
-                Type = Mutator.Arithmetic
-            };
-
-            var folder = new FolderComposite { Name = "RootFolder", RelativePath = "src" };
-
-            for (var i = 1; i <= 2; i++)
-            {
-                var addedFolder = new FolderComposite { Name = $"{i}", RelativePath = $"src/{i}" };
-                folder.Add(addedFolder);
-
-                for (var y = 0; y <= 4; y++)
-                {
-                    var m = new Collection<Mutant>();
-                    addedFolder.Add(new FileLeaf()
-                    {
-                        Name = $"SomeFile{i}.cs",
-                        RelativePath = $"src/{i}/SomeFile{i}.cs",
-                        Mutants = m,
-                        SourceCode = "void M(){ int i = 0 + 8; }"
-                    });
-
-                    for (var z = 0; z <= 5; z++)
-                    {
-                        m.Add(new Mutant()
-                        {
-                            Id = duplicateMutant ? 2 : new Random().Next(1, 5000),
-                            ResultStatus = 100 / 6 * z < mutationScore ? MutantStatus.Killed : MutantStatus.Survived,
-                            Mutation = mutation
-                        });
-                    }
-                }
-            }
-
-            return folder;
         }
     }
 }
